+++ src/utils/fs.js
@@ -1,0 +1,9 @@
+const fs = require('fs');
+
+function ensureDirSync(dirPath) {
+    if (!fs.existsSync(dirPath)) {
+        fs.mkdirSync(dirPath, { recursive: true });
+    }
+}
+
+module.exports = { ensureDirSync };