+++ test/suite/extension.test.js
@@ -12,4 +12,9 @@ suite('Extension Test Suite', () => {
 		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
 		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
 	});
+
+	test('Contributes addFeature command', async () => {
+		const cmds = await vscode.commands.getCommands(true);
+		assert.ok(cmds.includes('scratchtools.addFeature'));
+	});
 });